requirements.txt
streamlit
easyocr
pillow
torch
torchvision
transformers
opencv-python
